+++ .changeset/slimy-mangos-brake.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/atst': minor
+---
+
+Deprecate parseAttestationBytes and createRawKey in favor for createKey, createValue
