+++ packages/atst/src/lib/getEvents.ts
@@ -4,7 +4,7 @@ import { Address } from 'wagmi'
 import { ATTESTATION_STATION_ADDRESS } from '../constants/attestationStationAddress'
 import { abi } from '../lib/abi'
 import { AttestationCreatedEvent } from '../types/AttestationCreatedEvent'
-import { encodeRawKey } from './encodeRawKey'
+import { encodeRawKey } from './createKey'
 
 export const getEvents = async ({
   creator = null,
