+++ packages/atst/src/lib/prepareWriteAttestation.ts
@@ -4,7 +4,7 @@ import { formatBytes32String } from 'ethers/lib/utils.js'
 import { ATTESTATION_STATION_ADDRESS } from '../constants/attestationStationAddress'
 import { WagmiBytes } from '../types/WagmiBytes'
 import { abi } from './abi'
-import { stringifyAttestationBytes } from './stringifyAttestationBytes'
+import { createValue } from './createValue'
 
 export const prepareWriteAttestation = async (
   about: Address,
@@ -27,6 +27,6 @@ export const prepareWriteAttestation = async (
     abi,
     functionName: 'attest',
     chainId,
-    args: [about, formattedKey, stringifyAttestationBytes(value) as WagmiBytes],
+    args: [about, formattedKey, createValue(value) as WagmiBytes],
   })
 }
