+++ packages/atst/src/lib/prepareWriteAttestations.ts
@@ -4,7 +4,7 @@ import { formatBytes32String } from 'ethers/lib/utils.js'
 import { ATTESTATION_STATION_ADDRESS } from '../constants/attestationStationAddress'
 import { WagmiBytes } from '../types/WagmiBytes'
 import { abi } from './abi'
-import { stringifyAttestationBytes } from './stringifyAttestationBytes'
+import { createValue } from './createValue'
 
 type Attestation = {
   about: Address
@@ -27,7 +27,7 @@ export const prepareWriteAttestations = async (
         `key is longer than 32 bytes: ${attestation.key}.  Try using a shorter key or using 'encodeRawKey' to encode the key into 32 bytes first`
       )
     }
-    const formattedValue = stringifyAttestationBytes(
+    const formattedValue = createValue(
       attestation.value
     ) as WagmiBytes
     return {
